TimeScale fix for level one + compile error fix
This fixes the time control buttons for level 1 including the
highlighting for the fast forward button.  I also fixed the compile
error for the "Tutorial Button" by adding a temporary "fix" button
outside of the main camera inside level 1 scene.  This must be there for
now because the Object "Tutorial Button" is part of a script that is
also used in the main menu scene.  You can NOT delete, comment or make
this object private otherwise it will cause compile errors when working
in the main menu scene.