Version 2 (02/18)
David's second project version

Includes:

Changes to level 1 time control buttons.  The play button now sets
timeScale to 1 for all conditions.  The pause button only pauses and can
not be toggled to resume previous timeScale(Use 'Play' button instead).

Fixes lag issues for level 1 related to the 2D rigidbody collider and
the timeScale.  The fix was made in the script 'CircleEdgeCollider2D'
attached to the Cell Membrane object.  The script now has a conditional
check inside the System 'Update' method to stop calls to the method
'DrawCircle' which makes memory tasking physics calculations.  The
conditional check stops calls to 'DrawCircle' if the game is paused.
This works fine because the method 'DrawCircle' is called 'OnStart' for
initialization and is not needed to be called overy clock cycle when the
timeScale is '0' (paused).